Commit
Scenario:

The year was 2018 and the FIFA World Cup that year was hosted by Russia. The World Cup 2018 squad dataset was handed over to explore. As a Data Analyst in the Analytics team of a Football Academy company, my task was to uncover some hidden insights to help optimize the business's scouting decisions.

The Director of Football Analytics would like to know the extent of insights that can be uncovered from the dataset and make recommendations to improve and optimize the scouting strategies of the firm. The dataset can be found here and you can follow along to explore the data.

Objectives:

- What hypothesis can be formed from this data?

- Is there any attribute in this data that has any effect on players profile? (caps and number of goals).

- Hypothetically, it is assumed that age has an effect on players profile (in terms of caps and number of goals).

- Hypothesis: Does the hypothesis hold true that age has effect on players profile (caps and no of goals)?

Some likely questions that might interest the stakeholders from this data:

- Is there any correlation between 'goals', 'age' and 'caps'?

- Can we get the names of the countries belonging to each group?

- Is there any relationship between players position and number of goals scored?

- Which group has the highest number of goals?

- Who are the 10 most-capped players in the FIFA 2018 World Cup?

- Which of the clubs has the highest number of players represented at the 2018 world cup?

- Which country scored the highest number of goals at world cup tournaments to date as at world cup 2018?

- Who are the top 5 best players in the world as at 2018, based on number of goals scored by their countries in world cup tournaments to date?

- For scouting purpose, identify 5 best players per position based on goal-output and/or number of caps